+++ examples/viewer/src/App.jsx
@@ -21,7 +21,7 @@ import {
   LINK_TYPE,
 } from 'wix-rich-content-plugin-link/dist/module.viewer';
 import { imageTypeMapper } from 'wix-rich-content-plugin-image/dist/module.viewer';
-import { mapTypeMapper } from 'wix-rich-content-plugin-map/dist/module.viewer';
+import { mapTypeMapper, MAP_TYPE } from 'wix-rich-content-plugin-map/dist/module.viewer';
 import { Strategy as HashTagStrategy, Component as HashTag } from 'wix-rich-content-plugin-hashtag';
 import {
   createHeadersMarkdownDecorator,
@@ -32,6 +32,8 @@ import {
   MENTION_TYPE,
   mentionsTypeMapper,
 } from 'wix-rich-content-plugin-mentions/dist/module.viewer';
+import { fileUploadTypeMapper } from 'wix-rich-content-plugin-file-upload/dist/module.viewer';
+import { giphyTypeMapper } from 'wix-rich-content-plugin-giphy/dist/module.viewer';
 
 import 'wix-rich-content-common/dist/styles.min.css';
 import 'wix-rich-content-viewer/dist/styles.min.css';
@@ -46,6 +48,8 @@ import 'wix-rich-content-plugin-mentions/dist/styles.min.css';
 import 'wix-rich-content-plugin-video/dist/styles.min.css';
 import 'wix-rich-content-plugin-sound-cloud/dist/styles.min.css';
 import 'wix-rich-content-plugin-map/dist/styles.min.css';
+import 'wix-rich-content-plugin-file-upload/dist/styles.min.css';
+import 'wix-rich-content-plugin-giphy/dist/styles.min.css';
 
 import TestData from './TestData/initial-state';
 import styles from './App.scss';
@@ -92,6 +96,8 @@ class App extends Component {
       mentionsTypeMapper,
       imageTypeMapper,
       mapTypeMapper,
+      fileUploadTypeMapper,
+      giphyTypeMapper,
     ];
 
     this.config = {
@@ -103,6 +109,27 @@ class App extends Component {
       },
       [LINK_TYPE]: linkPluginSettings,
       [MENTION_TYPE]: mentionsPluginSettings,
+      [MAP_TYPE]: {
+        googleMapApiKey: process.env.GOOGLE_MAPS_API_KEY,
+        minWidth: 100,
+        maxWidth: 1400,
+        minHeight: 100,
+        maxHeight: 1400,
+        width: 650,
+        mapSettings: {
+          address: 'Wix HQ, Nemal Tel Aviv Street, Tel Aviv-Yafo, Israel',
+          locationDisplayName: 'Wix HQ, Nemal Tel Aviv Street, Tel Aviv-Yafo, Israel',
+          lat: 32.097235,
+          lng: 34.77427,
+          zoom: 18,
+          mode: 'roadmap',
+          isMarkerShown: true,
+          isZoomControlShown: true,
+          isStreetViewControlShown: true,
+          isViewControlShown: true,
+          isDraggingAllowed: true,
+        },
+      },
     };
 
     this.decorators = [