+++ packages/editor/src/RichContentEditor/RichContentEditor.jsx
@@ -21,6 +21,7 @@ import {
   normalizeInitialState,
   TooltipHost,
   TOOLBARS,
+  Context,
 } from 'wix-rich-content-common';
 import styles from '../../statics/styles/rich-content-editor.scss';
 import draftStyles from '../../statics/styles/draft.scss';
@@ -45,12 +46,13 @@ class RichContentEditor extends Component {
     this.initPlugins();
   }
 
+  setEditorState = editorState => this.setState({ editorState });
+
   initPlugins() {
     const { helpers, plugins, config, isMobile, anchorTarget, relValue, t } = this.props;
 
     const { theme } = this.state;
     const getEditorState = () => this.state.editorState;
-    const setEditorState = editorState => this.setState({ editorState });
     const { pluginInstances, pluginButtons, pluginTextButtons, pubsubs } = createPlugins({
       plugins,
       config,
@@ -61,7 +63,7 @@ class RichContentEditor extends Component {
       anchorTarget,
       relValue,
       getEditorState,
-      setEditorState,
+      setEditorState: this.setEditorState,
     });
     this.initEditorToolbars(pluginButtons, pluginTextButtons);
     this.pluginKeyBindings = initPluginKeyBindings(pluginTextButtons);
@@ -298,9 +300,9 @@ class RichContentEditor extends Component {
     );
   };
 
-  renderAccessibilityListener = () => <AccessibilityListener isMobile={this.props.isMobile} />;
+  renderAccessibilityListener = () => <AccessibilityListener />;
 
-  renderTooltipHost = () => <TooltipHost theme={this.state.theme} />;
+  renderTooltipHost = () => <TooltipHost />;
 
   render() {
     const { isMobile } = this.props;
@@ -310,19 +312,21 @@ class RichContentEditor extends Component {
       [theme.desktop]: !isMobile && theme && theme.desktop,
     });
     return (
-      <Measure bounds onResize={({ bounds }) => this.updateBounds(bounds)}>
-        {({ measureRef }) => (
-          <div style={this.props.style} ref={measureRef} className={wrapperClassName}>
-            <div className={classNames(styles.editor, theme.editor)}>
-              {this.renderAccessibilityListener()}
-              {this.renderEditor()}
-              {this.renderToolbars()}
-              {this.renderInlineModals()}
-              {this.renderTooltipHost()}
+      <Context.Provider theme={theme} {...this.props}>
+        <Measure bounds onResize={({ bounds }) => this.updateBounds(bounds)}>
+          {({ measureRef }) => (
+            <div style={this.props.style} ref={measureRef} className={wrapperClassName}>
+              <div className={classNames(styles.editor, theme.editor)}>
+                {this.renderAccessibilityListener()}
+                {this.renderEditor()}
+                {this.renderToolbars()}
+                {this.renderInlineModals()}
+                {this.renderTooltipHost()}
+              </div>
             </div>
-          </div>
-        )}
-      </Measure>
+          )}
+        </Measure>
+      </Context.Provider>
     );
   }
 }